backend/pyproject.toml
[tool.poetry]
name = "nestchat"
version = "0.1.0"
description = "Nestchat Backend"
authors = ["masingbackstage <filipswiderek@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"  
djangorestframework = "^3.15.0"
django-environ = "^0.12.0"
psycopg = "^3.2.10"
django-cors-headers = "^4.9.0"
django-filter = "24.2"
djangorestframework-camel-case = "1.4.2"

djangorestframework-simplejwt = "^5.5.1"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.12.1"

drf-spectacular = "^0.27.0" 

pillow = "^12.0.0"
django-storages = {extras = ["azure"], version = "^1.14.6"}

daphne = "^4.1"
channels-redis = "^4.2"

celery = "^5.5.3"
django-celery-beat = "^2.8.1"
redis = "^6.4.0"

livekit-api = "^0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
pytest-django = "^4.5"
isort = "^5.13"
black = "^24.0"
django-debug-toolbar = "^6.0.0"
watchfiles = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

